+++ crates/zng-view/src/lib.rs
@@ -176,12 +176,15 @@ use crate::{
 #[cfg(ipc)]
 zng_env::on_process_start!(|args| {
     if std::env::var("ZNG_VIEW_NO_INIT_START").is_err() {
-        if args.yield_count == 0 {
-            // give tracing handlers a chance to observe the view-process
-            return args.yield_once();
+        if !zng_env::about().is_test {
+            if args.yield_count == 0 {
+                // give tracing handlers a chance to observe the view-process
+                return args.yield_once();
+            }
+            view_process_main();
+        } else {
+            tracing::debug!("view-process not inited in test app");
         }
-
-        view_process_main();
     }
 });